Moving query cirterias into its own class
